+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/exceptions/UIRemoveAssociatedProjectException.java
@@ -1,0 +1,10 @@
+package ca.corefacility.bioinformatics.irida.ria.web.exceptions;
+
+/**
+ * Exception to through if there is an error removing an associated project from a project
+ */
+public class UIRemoveAssociatedProjectException extends Exception {
+	public UIRemoveAssociatedProjectException(String message) {
+		super(message);
+	}
+}