+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Changes
 * [Workflow]: Automated pipelines which become marked as disabled will now be prevented from running, similar to out-of-date automated pipelines.
 * [Developer]: Upgrade to Apache POI 5.2.2, and switched to excel-streaming-reader 3.6.1 from xlsx-streamer 2.1.0
 * [UI]: Added clarification that deleting or removing genomics data will not remove the underlying data files (e.g., sequencing data) but will only make it inaccessible.
+* [UI]: Updated project samples page to use Ant Design.
 
 22.01 to 22.03
 --------------